# docs/智能测试平台.openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "智能测试平台 API",
    "description": "智能测试平台完整接口文档，包含认证、项目管理、需求管理、测试点管理、测试用例管理、自动化执行六大模块",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "本地开发服务器"
    }
  ],
  "paths": {
...
    "/api/test-points": {
      "get": {
        "summary": "获取所有测试点",
        "description": "获取所有测试点列表",
        "operationId": "getAllTestPoints",
        "tags": ["测试点管理模块"],
        "security": [{"BasicAuth": []}],
        "responses": {
          "200": {
            "description": "成功获取测试点列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TestPoint"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "创建测试点",
        "description": "创建新的测试点",
        "operationId": "createTestPoint",
        "tags": ["测试点管理模块"],
        "security": [{"BasicAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTestPointRequest"
              },
              "example": {
                "projectId": 1,
                "requirementId": 1,
                "pointName": "登录功能测试点",
                "pointType": "FUNCTIONAL",
                "function": "用户登录",
                "sceneType": "NORMAL",
                "riskLevel": "HIGH",
                "businessRule": "用户必须使用正确的账号密码登录",
                "description": "验证用户登录功能",
                "automationSuggested": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestPoint"
                }
              }
            }
          }
        }
      }
    },
    "/api/test-points/requirement/{requirementId}": {
      "get": {
        "summary": "按需求获取测试点",
        "description": "获取指定需求下的所有测试点",
        "operationId": "getTestPointsByRequirementId",
        "tags": ["测试点管理模块"],
        "security": [{"BasicAuth": []}],
        "parameters": [
          {
            "name": "requirementId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "需求ID"
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取测试点列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TestPoint"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/test-points/project/{projectId}": {
      "get": {
        "summary": "按项目获取测试点",
        "description": "获取指定项目下的所有测试点",
        "operationId": "getTestPointsByProjectId",
        "tags": ["测试点管理模块"],
        "security": [{"BasicAuth": []}],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "项目ID"
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取测试点列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TestPoint"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/test-points/{id}": {
      "get": {
        "summary": "获取测试点详情",
        "description": "根据ID获取测试点详细信息",
        "operationId": "getTestPointById",
        "tags": ["测试点管理模块"],
        "security": [{"BasicAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "测试点ID"
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取测试点详情",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestPoint"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "更新测试点",
        "description": "更新测试点信息",
        "operationId": "updateTestPoint",
        "tags": ["测试点管理模块"],
        "security": [{"BasicAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "测试点ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTestPointRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestPoint"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "删除测试点",
        "description": "删除指定测试点",
        "operationId": "deleteTestPoint",
        "tags": ["测试点管理模块"],
        "security": [{"BasicAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "测试点ID"
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          }
        }
      }
    },
...
    "/api/test-cases/test-point/{testPointId}": {
      "get": {
        "summary": "按测试点获取测试用例",
        "description": "获取指定测试点下的所有测试用例",
        "operationId": "getTestCasesByTestPointId",
        "tags": ["测试用例管理模块"],
        "security": [{"BasicAuth": []}],
        "parameters": [
          {
            "name": "testPointId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "测试点ID"
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取测试用例列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TestCase"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TestPoint": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "测试点ID"
          },
          "projectId": {
            "type": "integer",
            "format": "int64",
            "description": "所属项目ID"
          },
          "requirementId": {
            "type": "integer",
            "format": "int64",
            "description": "所属需求ID"
          },
          "requirementItemId": {
            "type": "integer",
            "format": "int64",
            "description": "需求项ID"
          },
          "pointCode": {
            "type": "string",
            "description": "测试点编码"
          },
          "pointName": {
            "type": "string",
            "description": "测试点名称"
          },
          "pointType": {
            "type": "string",
            "description": "测试点类型"
          },
          "name": {
            "type": "string",
            "description": "名称"
          },
          "function": {
            "type": "string",
            "description": "功能描述"
          },
          "sceneType": {
            "type": "string",
            "description": "场景类型"
          },
          "riskLevel": {
            "type": "string",
            "description": "风险等级"
          },
          "businessRule": {
            "type": "string",
            "description": "业务规则"
          },
          "description": {
            "type": "string",
            "description": "描述"
          },
          "automationSuggested": {
            "type": "boolean",
            "description": "是否建议自动化"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          },
          "createdBy": {
            "type": "integer",
            "format": "int64",
            "description": "创建人ID"
          }
        }
      },
      "CreateTestPointRequest": {
        "type": "object",
        "required": ["projectId", "requirementId", "pointName"],
        "properties": {
          "projectId": {
            "type": "integer",
            "format": "int64",
            "description": "所属项目ID"
          },
          "requirementId": {
            "type": "integer",
            "format": "int64",
            "description": "所属需求ID"
          },
          "requirementItemId": {
            "type": "integer",
            "format": "int64",
            "description": "需求项ID"
          },
          "pointName": {
            "type": "string",
            "description": "测试点名称"
          },
          "pointType": {
            "type": "string",
            "description": "测试点类型"
          },
          "function": {
            "type": "string",
            "description": "功能描述"
          },
          "sceneType": {
            "type": "string",
            "description": "场景类型"
          },
          "riskLevel": {
            "type": "string",
            "description": "风险等级"
          },
          "businessRule": {
            "type": "string",
            "description": "业务规则"
          },
          "description": {
            "type": "string",
            "description": "描述"
          },
          "automationSuggested": {
            "type": "boolean",
            "description": "是否建议自动化"
          }
        }
      },
      "TestCase": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "用例ID"
          },
          "projectId": {
            "type": "integer",
            "format": "int64",
            "description": "所属项目ID"
          },
          "testPointId": {
            "type": "integer",
            "format": "int64",
            "description": "所属测试点ID"
          },
          "caseType": {
            "type": "string",
            "enum": ["FUNCTIONAL", "API", "UI", "PERFORMANCE"],
            "description": "用例类型"
          },
          "caseNumber": {
            "type": "string",
            "description": "用例编号"
          },
          "caseCode": {
            "type": "string",
            "description": "用例编码"
          },
          "title": {
            "type": "string",
            "description": "用例标题"
          },
          "precondition": {
            "type": "string",
            "description": "前置条件"
          },
          "steps": {
            "type": "string",
            "description": "测试步骤"
          },
          "expectedResult": {
            "type": "string",
            "description": "预期结果"
          },
          "priority": {
            "type": "string",
            "enum": ["HIGH", "MEDIUM", "LOW"],
            "description": "优先级"
          },
          "automation": {
            "type": "boolean",
            "description": "是否自动化"
          },
          "status": {
            "type": "string",
            "enum": ["PENDING_REVIEW", "REVIEWED", "APPROVED", "REJECTED"],
            "description": "状态"
          },
          "creator": {
            "type": "string",
            "description": "创建人"
          },
          "createdBy": {
            "type": "integer",
            "format": "int64",
            "description": "创建人ID"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "更新时间"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BasicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Basic认证，格式为 `Basic Base64(username:password)`"
      }
    }
  },
  "security": [
    {
      "BasicAuth": []
    }
  ]
}